+++ sympy/external/mpmath.py
@@ -144,4 +144,34 @@ __all__ = [
     "to_int",
     "to_rational",
     "to_str",
+    "_giant_steps_mpmath",
 ]
+
+
+def _giant_steps_mpmath(start, target, n=2):
+    """
+    Return a list of integers ~=
+
+    [start, n*start, ..., target/n^2, target/n, target]
+
+    but conservatively rounded so that the quotient between two
+    successive elements is actually slightly less than n.
+
+    With n = 2, this describes suitable precision steps for a
+    quadratically convergent algorithm such as Newton's method;
+    with n = 3 steps for cubic convergence (Halley's method), etc.
+
+        >>> from sympy.polys.ring_series import _giant_steps_mpmath
+        >>> _giant_steps_mpmath(50,1000)
+        [66, 128, 253, 502, 1000]
+        >>> _giant_steps_mpmath(50,1000,4)
+        [65, 252, 1000]
+
+    """
+    # This function is copied from mpmath to avoid depending on mpmath
+    # internals. This implementation could possibly be improved for the
+    # particular purpose here though.
+    L = [target]
+    while L[-1] > start*n:
+        L = L + [L[-1]//n + 2]
+    return L[::-1]
