+++ sympy/polys/ring_series.py
@@ -48,6 +48,7 @@ from sympy.polys.polyerrors import DomainError
 from sympy.polys.monomials import (monomial_min, monomial_mul, monomial_div,
                                    monomial_ldiv)
 from sympy.external.gmpy import factorial as ifac
+from sympy.external.mpmath import _giant_steps_mpmath
 from sympy.core import PoleError, Function, Expr
 from sympy.core.numbers import Rational
 from sympy.core.intfunc import igcd
@@ -87,34 +88,6 @@ def _invert_monoms(p1):
     for mvi, cvi in zip(mv, cv):
         p[(deg - mvi[0],)] = cvi
     return p
-
-def _giant_steps_mpmath(start, target, n=2):
-    """
-    Return a list of integers ~=
-
-    [start, n*start, ..., target/n^2, target/n, target]
-
-    but conservatively rounded so that the quotient between two
-    successive elements is actually slightly less than n.
-
-    With n = 2, this describes suitable precision steps for a
-    quadratically convergent algorithm such as Newton's method;
-    with n = 3 steps for cubic convergence (Halley's method), etc.
-
-        >>> from sympy.polys.ring_series import _giant_steps_mpmath
-        >>> _giant_steps_mpmath(50,1000)
-        [66, 128, 253, 502, 1000]
-        >>> _giant_steps_mpmath(50,1000,4)
-        [65, 252, 1000]
-
-    """
-    # This function is copied from mpmath to avoid depending on mpmath
-    # internals. This implementation could possibly be improved for the
-    # particular purpose here though. purpose here though.
-    L = [target]
-    while L[-1] > start*n:
-        L = L + [L[-1]//n + 2]
-    return L[::-1]
 
 def _giant_steps(target):
     """Return a list of precision steps for the Newton's method"""
